Implement deletePost
This commit implements post deletion from the client.

- a `deletePost` method was added to `MyApiService` which defines the route DELETE:"posts/"
to make an API call to delete the Post from the backend

- `deletePostFromServer` method added to `com.example.poste.models.User` class, which makes the
API call to the server to delete a Post. It takes a Post model (defined in
`com.example.poste.models.Post) and a callback as parameters. Callback will be discussed below.

- `PostDeletionCallback` was added to `com.example.poste.calbacks`. Please review the code for
usage details; I've added extensive docstrings. Since API calls are made asynchronously, we cannot
depend on return values like we might in a standard Java application. A 'callback' informs the
calling code of asynchronous success or failure -- this way, we can decide what to do if a request
succeeds or fails without waiting on the response.

- `DebugUtils` has had some extra utilities added

- RetrofitClient was slightly modified to add a logger

TL;DR: To add post deletion to an activity:
1) Define a PostDeletionCallback in your activity, implementing `onSuccess` and `onError` methods.
2) Call `User.getUser().deletePostFromServer`, passing a Post model and the callback.
3) If the call succeeds, the `onSuccess` method will run. Otherwise, the `onError` method will run.
4) If you have questions, look at the `User.deletePostFromServer` docstring