Add content sharing to Poste from outside app
This implements the 'share to Poste' functionality we have been discussing, by allowing a user to select
'share' from another application and share the data to Poste.

Normally, the application opens from IntroActivity, which allows user to choose to Login or Register.
This will change when a user attempts to share to Poste.

Alternate control flow:

1) User clicks 'Share' on some content and selects to share to Poste
2) Application opens NewPostActivity
3) Application checks if user is logged in.
4a) If user is logged in, proceed with post creation, returning to dashboard when done
4b) If user is not logged in, redirect to login page; when logged in, return to NewPostActivity and
proceed with post creation

New Files:
- `NewPostActivity.java`: Represents the activity to create a new post (rather than editing) -- I wanted
to work on a fresh activity rather than modifying EditPostActivity and having to touch too much of the
code in one task. Functionally very similar to EditPostActivity, but there is also a spinner (dropdown
menu) that allows a user to select the folder to create the post in. TODO: We need to determine what
happens if a User attempts to creat a Post, but has no Folders. Options: Create a default 'Home' folder,
give an error, redirect to FolderCreation, etc.
- `activity_new_post.xml`: Layout for the above activity

Changes:
- `AndroidManifest.xml`: Added intent-filter to allow sharing content to NewPostActivity
- `DashboardActivity.java`: Implemented Create Post functionality. Now, when user elects to create a post from
this page, it starts NewPostActivity, but does not finish -- this way, when post creation is finished,
user is returned here
- `LoginActivity.java`: Added checks for 'getBooleanExtra("return")'. This extra information should be passed
to `LoginActivity` when we want `LoginActivity` to simply `finish()` and return to the previous
activity when done, rather than redirecting user to the Dashboard. This is useful when a user tries to
share content to Poste, but they are not logged in -- first, it will ask them to login, but then return
them to NewPostActivity (which called Login) rather than to the Dashboard. If this extra isn't passed
to LoginActivity, they will simply be forwarded to dashboard
- `CreatePost.java`: Modified API request class to match sent data format to that expected by backend
- `Folder.java`: Implemented `toString()`, which is used to display Folder title in NewPostActivity spinner / dropdown
- `User.java`: Removed some duplicate code and unused import, and cleans up updateFoldersAndPosts
- `ValidationUtils.java`: Added some documentation