Add some conversions / string values to enums
Kim -- This is kinda messy and probably defeats the purposes of using an enum
in the first place. I'm just so used to using constants rather than enums...
and this is just how I got this working with the backend the way it is. Originally
I wanted to ensure we sent a string so we didnt have to change the backend, but the
backend ended up requiring changes anyway so... idk. Oh well.

This is tested working with the backend changes I made a PR for back to your branch.

Feel free to take / leave / modify as you see fit. Not trying to step on toes, just
trying to assist, so if you want to ignore these PR please feel free.